[ARM] Fixed do_data_abort() to handle translation aborts generated in normal context.
On real hardware its possible that we get a translation fault when we are emulation on behalf of a normal VCPU i.e. when we are in normal context. We need to handle such transation faults just like Linux handling fault generated by kernel code trying to access user space memory using copy_to_user() or copy_from_user().

Signed-off-by: Anup Patel <anup@brainfault.org>