[TOOLS] Update DTC compiler to help us specify 64-bit values in DTS scripts.
Please note that the unit of storage in a DTB or .S generated from DTS script is a cell. A cell in DTC compiler is always 32-bit values.
For 64-bit machine, we had to specify 64-bit values as 2 cells which is difficult to understand and mantain.

This patch modifies DTC compiler to forcefully assume values specified as 0x..ULL as 64-bit values.

Signed-off-by: Anup Patel <anup@brainfault.org>