[CORE] NET: use can_receive if available and fix mbuf usage for netbridge
Used the can_receive() hook to prevent sending the packets to destination.
Fixed mbuf usage as per the semantic - sender doesn't free the mbuf, so destination
should free it, when same mbuf is broadcasted to multiple destinations increment the
refcounts of both mbuf as well as its external data buffer.

Signed-off-by: Anup Patel <anup@brainfault.org>