[CORE] Implement mutex using lowlevel waitqueue APIs.
mutex are not excatly binary semaphore. We have this notion of ownership in mutex so, both vmm_mutex_lock() and vmm_mutex_unlock() should be called from Orphan VCPU (or Thread) context which not the case with vmm_semaphore_up() and vmm_semaphore_down() operations. The vmm_semaphore_up() can be called from IRQ context or Orphan VCPU context or Normal VCPU context.

Signed-off-by: Anup Patel <anup@brainfault.org>