[CORE] Simplify timer events to enable on-stack creation of timer events.
This patch simplifies the timer events by removing the global list of all events.

We no longer need to keep track of all timer events, this means we don't need explicit create and destroy functions for timer events.

This in-turn makes timer events more light-weight and also allows it to be created on stack.

Signed-off-by: Anup Patel <anup@brainfault.org>