[ARM] arm32: Introduce VTLB zoning
Until now, we implemented VTLB as set of VTLB lines and each VTLB line as set of VTLB entries.
The VTLB line was addressed based on virtual address and the VTLB entries in VTLB line were associative.
This approach tried to implement VTLB similar to cpu cache but did not scale well and only thing that worked well is a VTLB with just one line.

This patch introduces a new way of implementing VTLB called VTLB zoning.

According to VTLB zoning, we divide a VCPU VTLB into zones and each VTLB zone is a set of VTLB entiries.
The VTLB zoning is a heuristic based approach and each VTLB zone represents specific type of VTLB entires or VTLB pages.
In the new approach, we have total 350 entries divided into 3 zones: global (200 pages or entires), non-global (125 pages or entires), and virtual (25 pages or entires).
From various trials it was observed that for good performance the ratio of global entries to non-global entries should be as close to 8:5 (i.e. golden ratio).

The new VTLB approach further improves guest linux boot-time, CPU performance and memory read/write bandwidth.

Signed-off-by: Anup Patel <anup@brainfault.org>