[EMULATORS] GIC: Fixes to make GICC_PMR emulation close to real HW
According to GIC spec, interrupts with priority higher than value in GICC_PMR (or priority mask) register are passed to CPU.

Further GICv1 and GICv2 specs (as found on the Cortex-A9 and newer) states that the GICC_PMR reset value is zero.
This differs from the 0xf0 reset value used on 11MPCore.

Signed-off-by: Anup Patel <anup@brainfault.org>