[TESTS] Reduce arm32 test binary and relocation sizes
We reduce the arm32-test binary size by adjusting the sections so that all non-allocated sections go at the end.
To reduce the amount of data to be copied while relocation, we only copy the allocated regions of the binary.
As a result of the above two, we need to manually zero-out bss section as it no longer is part of the binary.

In addition, we also zero-out heap section to ensure we always get zeroed memory from basic test heap.

The end result is:-
1) Smaller binary (in current case size reducing by more than half)
 - Faster static patching
 - Faster flashing/copy of binary to boot source (Flash/SD-card)
2) Faster relocation

Also, the new relocation code takes care of overlapping regions (in memmove style)

Signed-off-by: Anup Patel <anup@brainfault.org>