[CORE] Improve host irq to support custom handler for each host irq
This patch adds support for customized PIC programming sequence for each host irq.

This means we have to specify handlers for each host irq.

The handler for a host irq will be based on nature of host irq or underlying PIC.
For e.g. For VIC all interrupts are to be handled as level-triggered whereas for GIC all interrupts require only eoi.

Signed-off-by: Anup Patel <anup@brainfault.org>