[CORE] Introduce cpuinit sections for init functions called on multiple CPUs of SMP system
This patch introduces a new cpuint section and corresponding __cpuinit function attribute in vmm_compiler.h

For UP systems, __cpuinit is same as __init

For SMP systems, __cpuinit is similar to __init but not same. All __init functions will freed once used but __cpuinit functions will not be freed untill all CPUs execute it.

The introduction of __cpuinit also improves code readability because init functions marked with __cpuinit tell us that this function is to be called on all CPUs of SMP systems.

Signed-off-by: Anup Patel <anup@brainfault.org>