[CORE] Fix ordering of setting started flag in vmm_timer_start()
In the last part of vmm_timer_start(), we used to program the clockchip for minimum duration and set tlcp->started flag to TRUE.

This way of setting the tlcp->started flag is buggy because it is quite possible that min_delta_ns for a clockchip is really small and we might get first clockchip interrupt before setting tlcp->started flag. If this happens then tlcp->started will never be set causing vmm_timer_schedule_next_event() to do nothing. All-in-all in such a situation we stop getting clockchip interrupts and system gets stuck.

This patch fixes the above described issue by the setting tlcp->started flag before programming the clockchip in vmm_timer_start().

Signed-off-by: Anup Patel <anup@brainfault.org>