[LIBS] netstack: Deprecate UIP network stack
uIP (micor IP) is a very primitive network stack.

We had done lot of changes to uIP get is working but, the socket APIs of uIP are very primitive.

Further uIP is not tailored for multi-threaded applications.

For all our purposes, lwIP suits well and lwIP is highly-portable and proven with many projects using it.

Signed-off-by: Anup Patel <anup@brainfault.org>