[CORE] net: Implement zero-copy in lwip transmit path
We implement zero-copy in the transmit/output path of lwip by converting the
lwip pbuf chain to mbuf chain.

Freeing this mbuf external data needs to be done by calling pbuf_free on only
the first of the pbuf pointers. The rest of the extdata derived from the pbuf
chain use a M_EXT_DONTFREE flag to avoid any kind of extdata-free routine.

Signed-off-by: Anup Patel <anup@brainfault.org>