[CORE] Introduce notifier APIs for creating notifier chains
Different parts of Xvisor might require a mechanism for sending notifications for interested clients.
For e.g. VFS will be interested in block device register/unregister events to implement auto-unmount a block device that is unregistered

Our implementation of notifier APIs is highly inspired from Linux notifier APIs so that its easy to understand and use.

Signed-off-by: Anup Patel <anup@brainfault.org>