[CORE] Add vmm_devtree_clock_frequency() and vmm_devtree_irq_xxx() APIs
This patch enriches vmm_devtree APIs so that we can avoid hard-coded strings in arch/ and drivers/ for reading irq numbers from device tree nodes.

The patch also adds vmm_devtree_clock_frequency() API which is a helper API for reading clock-frequency attribute of device tree node. This API is not related to arch_clk_xxx() APIs in anyway and it is just a helper API.

Lot of DTS files, arch code, drivers, and emulators get affected by this patch so all possible ARM host have been re-tested with this patch.

Signed-off-by: Anup Patel <anup@brainfault.org>