[EMULATORS] Use guest->vcpu_count for number of VCPUs in a guest instance
The GIC emulator and MPCore emulators were depending on device tree node attributes for determining the number of VCPUs in a guest instance, which is not the best way to do it.

To get number of VCPUs in a guest instance we should use vcpu_count member of guest instance (i.e. guest->vcpu_count)

This patch updates emulators to use guest->vcpu_count.

Signed-off-by: Anup Patel <anup@brainfault.org>