[ARM] arm32: Introduce dynamically assigned contextid for L1 tables
This patch introduces dynamically assigned (or on-the-fly incremented) contextid for level1 tables.

Every newly created level1 table will be assigned a new contextid. The lower 8-bits of this contextid will be used as ASID when switching translation tables.

We also revert the "[ARM] arm32: Fix hard-coded weight in arch_delay_loop_cycles()" change because this patch makes it redundant on Beagleboard-xM and Hackberry.

Signed-off-by: Anup Patel <anup@brainfault.org>