[ARM] Move bss section towards end of xvisor arm binary
The bss section is a non-loadable section which means it does not
contain any data to be loaded hence keeping this section between
loadable sections will only increase xvisor binary size.

In fact, all non-loadable sections such as (bss and stack) should
be towards the end of the xvisor binary.

The patch moves bss section towards end of xvisor binary.

Signed-off-by: Anup Patel <anup@brainfault.org>
Signed-off-by: Sukanto Ghosh <sukantoghosh@gmail.com>