[ARCH] Place the built-in dtb in a separate .devtree section
Placing the built-in dtb in a separate section allows control on
where to place it.

For ARM, we place it inside init region so that this space can be
reclaimed later. We also reserve some space for future use.

For x86, we don't treat it in a special manner. Maybe someday
later it will be placed in init region for x86 too.

Signed-off-by: Sukanto Ghosh <sukantoghosh@gmail.com>
Signed-off-by: Anup Patel <anup@brainfault.org>