[x86] Populate device tree from ACPI tables.
This patch populates the device tree from the ACPI tables.
This happens in early boot after host address space is setup.
After this all drivers don't need to depend on ACPI to query
device details. Every property of device is available in the
device tree.

Currently APIC and HPET are supported but code is generic enough
to handle other kinds.

Signed-off-by: Himanshu Chauhan <hschauhan@nulltrace.org>