[ARM] arm32ve: Make d-cache operations by set/way SMP safe.
A arm32ve VCPU may be preempted when doing d-cache operations
by set/way and this VCPU might get scheduled on another host CPU
on a SMP host.

To handle this situation we trap d-cache operations by set/way
for arm32ve VCPU and flush the entire d-cache when scheduled
on another host CPU. We use per-VCPU cpumask to keep track of
host CPUs on which d-cache flush is pending and we fill this
cpumask with ones when we see a d-cache operation by set/way.

Signed-off-by: Anup Patel <anup@brainfault.org>