[ARM] mmu_lpae: Make _setup_initial_ttbl() suitable for both arm32ve and arm64
The aarch64 compiler treats "&a" as address of global variable "a"
relative to current PC. In other words, "&a" for global variable "a"
mean PC + relative offset of "a". This way of treating "&a" is good
but breaks _setup_intial_ttbl() which assumes that "&a" is link time
absolute address of global variable "a".

This patch fixes above described issue in _setup_initial_ttbl() so
that it is suitable for both arm32ve and arm64.

Signed-off-by: Anup Patel <anup@brainfault.org>
Signed-off-by: Sukanto Ghosh <sukantoghosh@gmail.com>