[ARM] Introduce virt-v8 guest template for arm64 guest
There are no real development boards openly available for arm64
hence we introduce a completely paravirtualized virt-v8 guest
for folks interested in having a dummy or paravirtualized
hardware for guest.

This virt-v8 guest is a virtual machine with VirtIO-based
devices, GIC and ARMv8 generic timer.

This patch adds guest template for virt-v8 guest.

Signed-off-by: Sukanto Ghosh <sukantoghosh@gmail.com>