[ARM] Introduce generic armv8 board
With the new arm64 (or armv8) port we try to implement unified
Xvisor binary for all arm64 boards.

To achieve this we introduce generic armv8 board which is fully
device tree based (i.e. DTB-based) similar to linux arm64 port.

The only constraint with generic armv8 board support is that
actual HW board on which Xvisor runs should have GICv2 or higher
as PIC.

Signed-off-by: Sukanto Ghosh <sukantoghosh@gmail.com>