[TESTS] Use Image instead of zImage for booting Linux on ARM32 guests
The AArch32 instruction set is not excatly same as ARMv7 instruction set
because few ARMv7 coprocessor instruction which were there for compatibility
with ARMv6 and ARMv5 are now removed in AArch32 instruction set.
For e.g. Drain write buffer operation 'mcr 15, 0, r0, cr7, cr10, {4}'

The above difference in AArch32 instruction set and ARMv7 instruction set
causes Linux boot to fail on ARM32 guest for Xvisor ARM64.

This patch updates documentation and basic firmware Makefiles for all ARM32
guests so that we use uncompressed Image for booting Linux instead of zImage.

This also improves Linux booting time on ARM32 guests because we don't have
zImage decompression any more.

Please note that we are not restricting users from using zImage. With
this patch we are only trying to have same boot process for ARM32 guests
on ARM32 and ARM64 hosts.

Signed-off-by: Anup Patel <anup@brainfault.org>