[ARM] Rename default bridge instance to br0
All supported ARM host have DTS files creating a default vmm_netswitch
named "netbridge0" which is nothing but a software bridge.

In Linux world, software bridges are typically named "br0", "br1"
and so on.

For easy in understanding, we will name vmm_netswitch instances with
software bridge policy as "br0", "br1", and so on.

As per above rationale, we rename "netbridge0" to "br0" in DTS files of
various ARM hosts.

Signed-off-by: Anup Patel <anup@brainfault.org>