[LIBS] list: Add support hlist (hash table list) APIs
This patch add linux-like hlist APIs for ease in implementing
hash table bucket lists.

hlist is just a double linked lists with a single pointer list head
so that single pointer list head can be used to point to a hash table
bucket list. Of course, hlist will be not as efficient as dlist because
we don't have O(1) access to hlist tail element.

Signed-off-by: Anup Patel <anup@brainfault.org>