[CORE] Make vmm_spinlocks for UP debugging friendly
Currently, we did not do any kind of locking for vmm_spinlocks on
UP host which caused vmm_spinlock bugs untraceable on UP host.

This patch makes vmm_spinlock APIs to update a software flag when
running on UP system. The state of the software flag in-turn helps
us debugging on UP host and also fixes behaviour of trylock() APIs
on UP host.

Signed-off-by: Anup Patel <anup@brainfault.org>