[COMMANDS] vfs: Introduce host_load and guest_load sub-commands
Currently, the "vfs load" sub-command only loaded file contents to a
specific location of host memory but this is not very useful.

In general, we would like to have a vfs sub-commands capable of loading
file contents to specific location of host memory or specific location
of guest memory.

This patch replaces "vfs load" sub-command with two sub-commands namely
"vfs host_load" and "vfs guest_load".

Signed-off-by: Anup Patel <anup@brainfault.org>