[CORE] Allow vmm_vcpu_irq_wait_timeout only when no irqs are pending
It can happen that someone assert irq to VCPU while the VCPU is trying to
do wait-for-interrupt. If this situation is not handled then a VCPU might
be paused even with irqs pending.

This patch fixes vmm_vcpu_irq_wait_timeout() so that 'execute_pending' is
ensured to zero before VCPU is paused.

Signed-off-by: Anup Patel <anup@brainfault.org>