[CORE] Handle emulated device endianness via device emulation framework
Currently, each device emulator was taking care for endianness of device
to show to guest. This is not wrong but leads to lot of duplicate code
accross device emulators hence this patch makes device endianness as
an attribute of "struct vmm_emulator" and device emulation framework
does appropriate endianness conversion as required.

Signed-off-by: Anup Patel <anup@brainfault.org>