[CORE] vmm_devemu: Associate endianess with both data and device emulator
Currently, we associate endianess with only device emulator which not
the complete story.

In future for ARM32VE and ARM64, we might have Guest VCPU running with
endianess different from Xvisor endianess. In this situation, the data
to/from Guest VCPU has an endiness associated with it.

This patch makes endianess handling in vmm_devemu more generic so that
it is easier for us to implement cross-endian virtualization for any
architecture.

Signed-off-by: Anup Patel <anup@brainfault.org>