[TESTS] Update README for atomthreads RTOS on Realview-PB-A8 guest
Recently, we moved to boot time loading of Guest images from a block
device (i.e. RAM disk or some other disk). We have updated most of
our documentation as-per the new way of loading Guest images but
forgot to update documentation of running atomthreads RTOS on the
Realview-PB-A8 guest.

This patch updates README for atomthreads RTOS on Realview-PB-A8 guest.

Signed-off-by: Anup Patel <anup@brainfault.org>