[CORE] net: Add hub or repeater policy based netswitch
The most simplest policy for a netswitch is hub or repeater policy.

According to hub or repeater policy, a packet coming to netswitch from
a "src" port is simply broadcasted to all other ports in netswitch.

This simple/braindead netswitch policy can be very useful in scenarios
where we have many network ports in host HW and we want to pair each
guest network port with a single host HW network port.

Signed-off-by: Anup Patel <anup@brainfault.org>