[ARM] arm32ve: arm64: First-cut arch_defterm_early_putc() for debugging
Many times in bringing-up Xvison on new ARM board we are stuck at a point
even before arch_defterm_init() hence we are totally in the dark if we
don't have HW debugger for the new ARM board.

To solve this issue we already have arch_defterm_early_putc() which is
called even before arch_defterm_init(). This way of debugging is currently
used by Xvisor x86_64 for early prints.

This patch implements arch_defterm_early_putc() for arm32ve and arm64
boards. Eventually we will also have it for arm32 boards.

Signed-off-by: Anup Patel <anup@brainfault.org>