[CORE] Always use buddy allocator based heap management
It has been a while since we are using buddy allocator based
heap managements. This is a very widely used memory allocator
in unix-like OSes (including Linux).

The main issue (and the only issue) with buddy allocator is
internal fragmentation because it always allocate memory in
power of 2 and some memory is always wasted in each allocation.
(Refer, http://en.wikipedia.org/wiki/Buddy_memory_allocation)

Most unix-like OSes have their own way of dealing with internal
fragmentation. Linux uses SLAB and SLUB allocator for this. We
currently have mempool APIs for this and in-future we will have
more methods of reducing internal fragmentation.

This patch makes buddy allocator mandatory for heap management.

Signed-off-by: Anup Patel <anup@brainfault.org>