[ARM] Rename versatile_clcd_setup() to versatile_clcd_setup_dma()
We will be soon having CLCD support for Versatile platform which
would require us to define separate versatile_clcd_setup() for
Versatile platform.

This patch rename versatile_clcd_setup() to versatile_clcd_setup_dma()
so that it is easy for us to define separate versatile_clcd_setup()
for Versatile platform.

Signed-off-by: Anup Patel <anup@brainfault.org>