[ARM] arm32: Minor improvement to domain fault and permission fault handling
To handle both domain fault and permission fault we typically determine
the faulting page and process it. For both type of faults if we are unable
to retrive faulting page then we simple remove the virtual address from VTLB
and resume the faulting instruction hoping that we will get translation fault
which in-turn will trigger appropriate fault to Guest VCPU.

This patch simplifies the above by directly calling translation fault handling
if we are unable to retrive faulting page in domain fault and permission fault
handling.

Signed-off-by: Anup Patel <anup@brainfault.org>