[DRIVERS] base: Adding base driver for probing pinctrl.
This patch adds:
Linux-like base driver for probing and extracting pinctrl settings
of any device driver before it's actual probe gets executed.
This patch implemets vmm_devdrv_pinctrl_bind() in drivers/base/pinctrl.c
which is defied as a weak function in core/vmm_devdrv.c

vmm_devdrv_pinctrl_bind() returns VMM_EPROBE_DEFER, which means some
device's driver is probed before the actual pinctrl driver (on which
this device's pinctrl settings are dependant) is not probed.
In this case we defer the probing of this new device so that master
pinctrl device driver can be probed first.

Signed-off-by: Pranavkumar Sawargaonkar <pranav.sawargaonkar@gmail.com>