[CORE] Add "bound_on" field for each vmm_clockchip
We need to track the host CPU to which a vmm_clockchip is bound
and we also need to ensure that a vmm_clockchip delivers interrupts
to host CPU it is bound to.

To achieve this we introduce "bound_on" field in vmm_clockchip
to represent host CPU bound to vmm_clockchip. When vmm_clockchip
is new or not bound to any host CPU, the "bound_on" is set to
UINT_MAX.

We also remove vmm_clockchip_find_best() API and instead introduce
two new APIs: vmm_clockchip_bind_best() and vmm_clockchip_unbind().

Signed-off-by: Anup Patel <anup@brainfault.org>