[ARM] Use alloced_rom and alloced_ram instead of rom and ram
Currently, we use reserved rom and ram regions for all types of
ARM guests hence we have to explicitly specify host physical
address of these resered rom and ram regions. This in-turn
makes guest creation process bit more complex then it should be.

This patch replaces reserved rom and ram regions with alloced
rom and ram regions. As a result of this patch, the guest
creation process for all types of ARM guests is much more
simplified with less commands.

Signed-off-by: Anup Patel <anup@brainfault.org>