[TESTS] Add fully paravirtualized virt-v7 guest for arm32ve
This patch adds basic firmware source and linux config files for virt-v7
guest.

The virt-v7 guest is excatly same as virt-v8 guest the only difference
it that former is arm32 guest whereas later is arm64 guest. Also, both
virt-v7 and virt-v8 are compatible to standardized mach-virt guest
which Linaro is pushing for KVM and Xen.

Signed-off-by: Anup Patel <anup@brainfault.org>