[ARM] arm32ve: arm64: Remove dependency of arch_regs.h on generic_timer.h
The arch_regs.h is indirectly included at potentially everywhere. We should
have minimum dependency of arch_regs.h on other headers hence this patch.

Also, this patch makes it clear that if we want a structure instance in
arch_regs.h then the structure should be defined in arch_regs.h or the
structure instance should be a void pointer.

Signed-off-by: Anup Patel <anup@brainfault.org>