[CORE] vmm_host_aspace: Reserve RAM areas as provided by arch code
This patch adds mechanism for arch code to provide reserved RAM areas
at host aspace init time. This can be very useful on SMP host where
secondary cores are spinning on some RAM area which should never be
overwritten. It is also useful when we have additional microcontroller
in SOC that runs a firmware residing in some part of RAM.

Signed-off-by: Anup Patel <anup@brainfault.org>