[ARM] generic_timer: Don't compute WFI timeout based on generic timer
It is not necessary that Guest VCPU will only wait for timer interrupt
when doing WFI. It might also wait for an IPI or some other interrupt.

This patch removes dependency of WFI timeout on generic timer state
by introducing timer events to keep track of physical timer and virtual
timer when Guest VCPU is not running.

Signed-off-by: Anup Patel <anup@brainfault.org>