[TOP] Improve dependency generation in Makefile
Currently, we generate and merge all built-in and module object
dependency files into one huge dependency file and then include
this huge dependency file.

The most important problem with the current approach is that
when someone accidently add object file for non-existent source
file then Make only complains that it was not able to generate
the huge dependency file. Instead of this Make should explicitly
tell us which object file was missing.

This patch removes the need of one huge dependency file and
re-organizes inclusion of dependency .dep files so that Make
shows correct complain when it is not able to compile some
object file.

With this improvements in Makefile the build process is also
much more faster now.

Signed-off-by: Anup Patel <anup@brainfault.org>