[ARM] arm32: arm32ve: Update delay loop to machine cycle conversion
As per latest Xvisor, the amount of machine cycles spent is twice
the number of delay loop iterations for arm32 and arm32ve. Previously,
this was thrice the number of delay loop iterations but it have
come back to expected normal value due to recent improvements such as
read-only pages for code sections. Let's hope that we dont have to
change it with every release.

Signed-off-by: Anup Patel <anup@brainfault.org>