[ARM] arm32: Don't use compile time check to override VTLB page shareability
With the increasing use of generic board support for all ports, it is
quite common that we endup running SMP-enabled Xvisor on UP boards
such as Hackberry or Beagleboard-xM.

The issue here is that on UP boards we only have local monitor and
global monitor is not present which causes SYNC data abort on UP board
if we do ldrex/strex on pages with shareability bit set.

In order to avoid above described issue, we remove the compile time
SMP check and instead use runtime SMP check to override VTLB page
shareablility bit.

Signed-off-by: Anup Patel <anup@brainfault.org>