[LIBS] Add support for a default user for libauth.
In case the filesystem is not present and libauth is
enabled, a default user can be added to the user. The
default username is admin. Password to supplied in menuconfig
is clear text. At compile time hash of the password
is generated and linked in the data.

Default password is EQuiN0X

Signed-off-by: Himanshu Chauhan <hschauhan@nulltrace.org>
Signed-off-by: Anup Patel <anup@brainfault.org>