[CORE] Avoid the "always_inline function might not be inlinable" warning.
GCC displays a warning about the fact a function may not be inlined even if
told so, if no optimization is asked for. To avoid this, as explained in the
GCC manual :
Generally, functions are not inlined unless optimization is specified. For
functions declared inline, this attribute inlines the function even if no
optimization level is specified.

Adding "inline" correct this warning.

Signed-off-by: Jimmy Durand Wesolowski <jimmy.durand-wesolowski@openwide.fr>