[CORE] Add config option for VA Pool aligment size
Buddy initialization for VA pool has the requirement
of aligning with the max bin selected for buddy.
As part of commit 82553c1a99dc75363b1c437eff8ece419dcef607
the max bins were increased to 23 for support 8MB. This
imposes VA pool alignment of 8MB.

x86 uses alignmen of 2MB and uses it as text start.

This patch makes this configuration a architecture specific
config option.

Signed-off-by: Himanshu Chauhan <hschauhan@nulltrace.org>